커뮤니티 파일 관련 코드 삭제 /  Product 파일 관련 코드 추가(주석으로 표시해 놓았음) /  mapper 패키지 resource로 이동(배포때문에 경로 변경)
