+++ src/main/java/com/koreait/dooboo/communityboard/controller/CommunityBoardController.java
@@ -1,23 +1,17 @@
 package com.koreait.dooboo.communityboard.controller;
 
-import java.util.ArrayList;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
-import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.servlet.ModelAndView;
 
 import com.koreait.dooboo.communityboard.command.BoardCommand;
 import com.koreait.dooboo.communityboard.dto.CommunityBoardDTO;
-import com.koreait.dooboo.product.dto.ProductimageDTO;
-import com.koreait.dooboo.util.FileUpload;
 
 @Controller
 /*@RequestMapping("/CommunityBoard")*/
@@ -74,27 +68,17 @@ public class CommunityBoardController {
 		
 		return mav;
 	}
-	 
+	
 	@PostMapping(value ="b.save")
 	@ResponseBody
-	public Map<String,Object> openBoardSave(CommunityBoardDTO boardDTO,@RequestParam("uploadFile") MultipartFile[] uploadFiles) {
-		
-		List<FileUpload> files = new ArrayList<FileUpload>(); 
-		if(null != uploadFiles) {
-			for(MultipartFile multiPartFile : uploadFiles) {
-				FileUpload file = new FileUpload(multiPartFile);
-				System.out.println(file.getFileName());
-				System.out.println(file.getOrgFileName());
-				files.add(file);
-			}
-			boardDTO.setFileUploadList(files);
-		}
+	public Map<String,Object> openBoardSave(CommunityBoardDTO boardDTO){
 		
 		boolean saveYn = boardCommand.registerBoard(boardDTO);
 		Map<String,Object> resultMap = new HashMap<String,Object>();
 		resultMap.put("saveYn", saveYn);
 		return resultMap;
 	}
+	 
 
 	@PostMapping(value ="b.delete")
 	@ResponseBody
@@ -104,14 +88,4 @@ public class CommunityBoardController {
 		resultMap.put("deleteYn", deleteYn);
 		return resultMap;
 	}
-	
-	@PostMapping(value ="deleteFile")
-	@ResponseBody
-	public Map<String,Object> setDeleteFile(ProductimageDTO productImageDTO) {
-		boolean deleteYn = boardCommand.setDeleteFile(productImageDTO);
-		Map<String,Object> resultMap = new HashMap<String,Object>();
-		resultMap.put("deleteYn", deleteYn);
-		return resultMap;
-	}
-
 }
