--- src/main/java/com/koreait/dooboo/mapper/communityboard.xml
@@ -1,178 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE mapper 
-PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" 
-"http://mybatis.org/dtd/mybatis-3-mapper.dtd" >
-
-<mapper namespace="com.koreait.dooboo.communityboard.dao.CommunityBoardDAO">
-
-<sql id="startPaging">
-  	 	SELECT
-	  	 	PAGE_B.*
-  	 	FROM(
-	  	 	SELECT
-		  	 	ROWNUM AS RNUM
-		  	 	,PAGE_A.*
-	  	 	FROM(  	 	
-</sql>
-<sql id="endPaging">
-			) PAGE_A
-			WHERE ROWNUM <![CDATA[<=]]> #{endPage}
-		) PAGE_B
-		WHERE RNUM <![CDATA[>=]]> #{startPage}
-</sql>
-<sql id="communityBoardColumns">
-		  BOARD_NO
-		, TITLE
-		, CONTENT
-		, HIT
-		, REG_NO
-		, MOD_NO
-		, REG_DATE
-		, MOD_DATE
-</sql>
-	<resultMap id="boardResult" type="com.koreait.dooboo.communityboard.dto.CommunityBoardDTO">
-	    <result property="rnum" column="RNUM"/>
-	    <result property="boardNo" column="BOARD_NO"/>
-	    <result property="title" column="TITLE"/>
-	    <result property="content" column="CONTENT"/>
-	    <result property="hit" column="HIT"/>
-	    <result property="regNo" column="REG_NO"/>
-	    <result property="modNo" column="MOD_NO"/>
-	    <result property="regDate" column="REG_DATE"/>
-	    <result property="modDate" column="MOD_DATE"/>
-	    <result property="nickname" column="NICKNAME"/>
-	    <collection property="productImageList" column="BOARD_NO" javaType="java.util.ArrayList" ofType="com.koreait.dooboo.product.dto.ProductimageDTO" select="selectBoardFileList"/>
-	</resultMap>
-
-	<insert id="insertBoard" parameterType="com.koreait.dooboo.communityboard.dto.CommunityBoardDTO">
-	INSERT INTO COMMUNITYBOARD (
-					  BOARD_NO
-					, TITLE
-					, CONTENT
-					, HIT
-					, REG_NO
-					, MOD_NO
-					, REG_DATE
-					, MOD_DATE
-		) VALUES (
-			  COMMUNITYBOARD_SEQ.NEXTVAL
-			, #{title}
-			, #{content}
-			, 0
-			, #{regNo}
-			, #{modNo}
-			, SYSDATE
-			, SYSDATE
-		)
-	</insert>
-
-	<select id="selectBoardDetail" parameterType="long" resultMap="boardResult">
-		SELECT
-			ROWNUM ,
-			COM.BOARD_NO ,
-			COM.TITLE ,
-			COM.CONTENT ,
-			COM.HIT ,
-			COM.REG_NO,
-			COM.MOD_NO,
-			COM.REG_DATE,
-			COM.MOD_DATE,
-			M.NICKNAME
-		FROM
-			COMMUNITYBOARD COM 
-			LEFT OUTER JOIN MEMBER M
-            ON REG_NO = M.MEMBER_NO
-		WHERE
-			BOARD_NO = #{board_no}
-	</select>
-
-	<update id="updateBoard" parameterType="com.koreait.dooboo.communityboard.dto.CommunityBoardDTO">
-		UPDATE COMMUNITYBOARD
-		SET
-			  MOD_DATE = SYSDATE
-			, TITLE = #{title}
-			, CONTENT = #{content}
-		WHERE
-			BOARD_NO = #{boardNo}
-	</update>
-
-	<update id="deleteBoard" parameterType="com.koreait.dooboo.communityboard.dto.CommunityBoardDTO">
-		UPDATE COMMUNITYBOARD
-		SET
-			  DEL_YN = 'Y'
-			, MOD_DATE = SYSDATE
-		WHERE
-			BOARD_NO = #{boardNo}
-	</update>
-	<select id="selectBoardFileList" resultType="com.koreait.dooboo.product.dto.ProductimageDTO">
-		SELECT * FROM PRODUCT_IMAGE WHERE PRODUCT_NO = #{productNo}
-	</select>
-	
-	<select id="selectBoardList" parameterType="com.koreait.dooboo.communityboard.dto.CommunityBoardDTO" resultMap="boardResult">
-	
-			
-  	 	<include refid="startPaging"/>
-	  		SELECT
-	  			ROWNUM,
-				COM.BOARD_NO ,
-				COM.TITLE ,
-				COM.CONTENT ,
-				COM.HIT ,
-				COM.REG_NO,
-				COM.MOD_NO,
-				COM.REG_DATE,
-				COM.MOD_DATE,
-				M.NICKNAME
-			FROM
-				COMMUNITYBOARD COM 
-			LEFT OUTER JOIN MEMBER M
-            ON REG_NO = M.MEMBER_NO
-			WHERE DEL_YN = 'N'
-			<if test="searchType != '' and searchType != null">
-				<if test="searchKeyword != '' and searchKeyword != null">
-					<if test="searchType == 'title' ">
-					 	AND TITLE like '%'|| #{searchKeyword} ||'%'
-					 </if>
-					<if test="searchType == 'content' ">
-					 	AND CONTENT like '%'|| #{searchKeyword} ||'%'
-					 </if>
-				</if>
-			</if>
-			
-			ORDER BY
-				COM.BOARD_NO DESC,
-				COM.REG_DATE DESC 
-		<include refid="endPaging"/>				
-	</select>
-	<update id="updateHit" parameterType="long">
-		UPDATE
-			COMMUNITYBOARD 
-		SET 
-			HIT = HIT + 1
-		WHERE
-			BOARD_NO = #{boardNo}
-	</update>
-
-	<select id="selectBoardTotalCount" parameterType="com.koreait.dooboo.communityboard.dto.CommunityBoardDTO" resultType="int">
-		SELECT
-			COUNT(*)
-		FROM
-			COMMUNITYBOARD
-		WHERE
-			del_yn = 'N'
-			<if test="searchType != '' and searchType != null">
-				<if test="searchKeyword != '' and searchKeyword != null">
-					<if test="searchType == 'title' ">
-					 	AND TITLE like '%'|| #{searchKeyword} ||'%'
-					 </if>
-					<if test="searchType == 'content' ">
-					 	AND CONTENT like '%'|| #{searchKeyword} ||'%'
-					 </if>
-				</if>
-			</if>
-	</select>
-	
-	<delete id="deleteFile" parameterType="com.koreait.dooboo.product.dto.ProductimageDTO">
-		DELETE FROM PRODUCT_IMAGE WHERE FILE_NO = #{fileNo} <!-- AND REG_NO = #{regNo} -->
-	</delete>	
-</mapper>
