--- src/main/java/com/koreait/dooboo/mapper/member.xml
@@ -1,66 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE mapper 
-PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" 
-"http://mybatis.org/dtd/mybatis-3-mapper.dtd" >
-
-<mapper namespace="com.koreait.dooboo.member.dao.MemberDAO">
-	<insert id="join" >
-		INSERT INTO MEMBER 
-		VALUES
-		(Member_SEQ.NEXTVAL,
-		#{name},
-		#{phone},
-		#{memberId},
-		#{password},
-		#{email},
-		#{nickname},
-		#{gender}, 
-		SYSDATE, 
-		null , 
-		0 , 
-		#{address}, 
-		#{birthday},
-		null,
-		#{apiNumber},
-		#{apiMemberNo}) <!-- status 0 = 정상   , 1= 삭제 -->
-	</insert>
-	<!-- 아이디와 비밀번호가 일치하고 status 가 0 인 회원 -->
-	<select id="login" resultType="com.koreait.dooboo.member.dto.MemberDTO">
-		SELECT * FROM MEMBER WHERE MEMBER_ID = #{memberId} AND PASSWORD= #{password} AND STATUS = 0
-	</select>
-	<update id="updateInfo">
-		UPDATE MEMBER SET NAME = #{param1} , NICKNAME = #{param2} , MOD_DATE = SYSDATE WHERE MEMBER_NO = #{param3} AND STATUS = 0
-	</update>
-
-	<select id="selectMemberByMemberNo" resultType="com.koreait.dooboo.member.dto.MemberDTO">
-		SELECT * FROM MEMBER WHERE MEMBER_NO = #{param1} AND STATUS = 0
-	</select>
-
-	
-	<select id="apiLoginCheck" resultType="com.koreait.dooboo.member.dto.MemberDTO">
-		SELECT * 
-		  FROM MEMBER
-		 WHERE APINUMBER = #{apiNumber} AND APIMEMBERNO = #{apiMemberNo}
-	</select>
-	
-	<update id="updatePassword">
-		UPDATE MEMBER SET PASSWORD = #{param1} , MOD_DATE = SYSDATE WHERE STATUS = 0 AND MEMBER_NO = #{param2}
-	</update>
-	<update id="updateContact">
-		UPDATE MEMBER SET PHONE = #{param1} , ADDRESS = #{param2} , MOD_DATE = SYSDATE WHERE STATUS = 0 AND MEMBER_NO = #{param3} 
-	</update>
-	<!-- 이름과 이메일에 해당하는 회원의 아이디 찾기 -->
-	<select id="findIdByEmailAndName" resultType="com.koreait.dooboo.member.dto.MemberDTO">
-		SELECT * FROM MEMBER WHERE NAME = #{name} AND EMAIL = #{email} AND STATUS = 0
-	</select>
-	<!-- 비밀번호를 변경하기 위해 , 아이디와 이메일에 해당하는 회원의 아이디 찾기 -->
-	<select id="selectMemberByEmailAndId" resultType="com.koreait.dooboo.member.dto.MemberDTO">
-		SELECT * FROM MEMBER WHERE MEMBER_ID = #{memberId} AND EMAIL = #{email} AND STATUS = 0
-	</select>
-	
-	<!-- 입력받은 아이디에 해당하는 회원이 있는지 조회 -->
-	<select id="searchByMemberId" resultType="int">
-		SELECT COUNT(MEMBER_ID) FROM MEMBER WHERE MEMBER_ID = #{param1}
-	</select>
-
-</mapper>
