--- src/main/java/com/koreait/dooboo/mapper/product.xml
@@ -1,144 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE mapper 
-PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" 
-"http://mybatis.org/dtd/mybatis-3-mapper.dtd" >
-
-<mapper namespace="com.koreait.dooboo.product.dao.ProductDAO">
-
-<sql id="startPaging">
-  	 	SELECT
-	  	 	PAGE_B.*
-  	 	FROM(
-	  	 	SELECT
-		  	 	ROWNUM AS RNUM
-		  	 	,PAGE_A.*
-	  	 	FROM(  	 	
-</sql>
-<sql id="endPaging">
-			) PAGE_A
-			WHERE ROWNUM <![CDATA[<=]]> #{endPage}
-		) PAGE_B
-		WHERE RNUM <![CDATA[>=]]> #{startPage}
-</sql>
-<sql id="productColumns">
-		  PRODUCT_NO
-		, TITLE
-		, PRICE
-		, HIT
-		, CONTENT
-		, STATUS
-		, ADDRESS
-		, REG_NO
-		, MOD_NO
-		, REG_DATE
-		, MOD_DATE
-	</sql>
-
-<select id="selectProductList" parameterType="com.koreait.dooboo.product.dto.ProductDTO" resultType="com.koreait.dooboo.product.dto.ProductDTO">
-	
-  	 	<include refid="startPaging"/>
-	  		SELECT
-	  			 ROWNUM
-				, P.PRODUCT_NO
-				, P.TITLE
-				, P.PRICE
-				, P.HIT
-				, P.CONTENT
-				, P.STATUS
-				, P.ADDRESS
-				, P.REG_NO
-				, P.MOD_NO
-				, P.REG_DATE
-				, P.MOD_DATE
-			FROM
-				PRODUCT P
-			LEFT OUTER JOIN MEMBER M
-            ON REG_NO = M.MEMBER_NO
-			WHERE DEL_YN = 'N'
-			ORDER BY
-				P.PRODUCT_NO DESC,
-				P.REG_DATE DESC 
-		<include refid="endPaging"/>				
-	</select>
-	
-	<update id="updateHit" parameterType="long">
-		UPDATE
-			PRODUCT 
-		SET 
-			HIT = HIT + 1
-		WHERE
-			PRODUCT_NO = #{productNo}
-	</update>
-	
-	<select id="selectProductTotalCount" parameterType="com.koreait.dooboo.product.dto.ProductDTO" resultType="int">
-		SELECT
-			COUNT(*)
-		FROM
-			PRODUCT
-		WHERE
-			del_yn = 'N'
-	</select>
-
-	<!--  @@@@@@@@@@@@@@@@@@ 업로드한 이미지 파일을 관리한다     보람씨 작업한것 @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@-->
-	<insert id="insertProductImage" parameterType="com.koreait.dooboo.product.dto.ProductimageDTO">
-		INSERT INTO PRODUCT_IMAGE
-		(	
-			 PRODUCT_NO
-			,FILE_NO
-			,FILE_PATH
-			,FILE_NAME
-			,REG_NO
-			,REG_DATE
-		)VALUES(
-			  #{productNo}
-			, PRODUCTIMAGE_SEQ.NEXTVAL 
-			, #{filePath} 
-			, #{fileName} 
-			, #{regNo} 
-			, SYSDATE
-		)
-	</insert>
-<!-- @@@@@@@@@@@@@@@@@@@@@@@@@@@@@보람씨 작업한것@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@ -->
-
-	
-	
-	
-	
-	<!-- 온석태 수정 -->
-		<!-- 판매상품을 등록한다 -->
-	<insert id="insertSellProduct">
-		INSERT INTO PRODUCT VALUES
-		(PRODUCT_SEQ.NEXTVAL,#{title},${price},0,#{content},1,#{address},#{regNo},null,SYSDATE,null)
-	</insert>
-	
-	<select id="getOneLocation" parameterType="long" resultType="com.koreait.dooboo.map.dto.MapDTO">
-		SELECT *
-		  FROM MAP
-		 WHERE MAP_NO = #{mapNo}
-	</select>
-	<select id="getProductNo" resultType="com.koreait.dooboo.product.dto.ProductDTO">
-		SELECT A.RN ,A.PRODUCT_NO
-		  FROM (SELECT ROWNUM AS RN,PRODUCT_NO,REG_NO
-		  		  FROM PRODUCT
-		  		 ORDER BY REG_DATE DESC)A
-		 WHERE A.RN = 1 AND A.REG_NO = #{regNo}  		 
-	</select>
-	
-	<select id="selectOneProduct" resultType="com.koreait.dooboo.product.dto.ProductDTO">
-		SELECT * 
-		  FROM PRODUCT 
-		 WHERE PRODUCT_NO = #{param1}
-	</select>
-	
-	<select id="getLikesCount" resultType="int">
-		SELECT COUNT(*)
-		  FROM LIKEPRODUCT
-		 WHERE PRODUCT_NO = #{productNo}
-	</select>
-	
-	<insert id="insertLikeProduct" parameterType="com.koreait.dooboo.likeproduct.dto.LikeProductDTO">
-		INSERT INTO LIKEPRODUCT
-		VALUES(#{productNo},#{regNo},SYSDATE)
-	</insert>
-
-</mapper>
