+++ src/main/java/com/koreait/dooboo/product/dao/ProductDAO.java
@@ -6,7 +6,12 @@ import com.koreait.dooboo.product.dto.ProductimageDTO;
 
 public interface ProductDAO {
 	//보람씨가 쓸 DAO
-	public int insertProductImage(ProductimageDTO productimageDTO); 
+	public int insertProductImage(ProductimageDTO productimageDTO);
+	
+	public int deleteFile(ProductimageDTO productImageDto);
+
+	public int updateBoard(ProductDTO params);
+	
 	//----------------------------------------------------------------
 	
 	public MapDTO getOneLocation(long mapNo);
