+++ src/main/resources/mapper/communityboard.xml
@@ -30,20 +30,7 @@ PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
 		, REG_DATE
 		, MOD_DATE
 </sql>
-	<resultMap id="boardResult" type="com.koreait.dooboo.communityboard.dto.CommunityBoardDTO">
-	    <result property="rnum" column="RNUM"/>
-	    <result property="boardNo" column="BOARD_NO"/>
-	    <result property="title" column="TITLE"/>
-	    <result property="content" column="CONTENT"/>
-	    <result property="hit" column="HIT"/>
-	    <result property="regNo" column="REG_NO"/>
-	    <result property="modNo" column="MOD_NO"/>
-	    <result property="regDate" column="REG_DATE"/>
-	    <result property="modDate" column="MOD_DATE"/>
-	    <result property="nickname" column="NICKNAME"/>
-	    <collection property="productImageList" column="BOARD_NO" javaType="java.util.ArrayList" ofType="com.koreait.dooboo.product.dto.ProductimageDTO" select="selectBoardFileList"/>
-	</resultMap>
-
+	
 	<insert id="insertBoard" parameterType="com.koreait.dooboo.communityboard.dto.CommunityBoardDTO">
 	INSERT INTO COMMUNITYBOARD (
 					  BOARD_NO
@@ -66,7 +53,7 @@ PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
 		)
 	</insert>
 
-	<select id="selectBoardDetail" parameterType="long" resultMap="boardResult">
+	<select id="selectBoardDetail" parameterType="long" resultType="com.koreait.dooboo.communityboard.dto.CommunityBoardDTO">
 		SELECT
 			ROWNUM ,
 			COM.BOARD_NO ,
@@ -104,13 +91,10 @@ PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
 		WHERE
 			BOARD_NO = #{boardNo}
 	</update>
-	<select id="selectBoardFileList" resultType="com.koreait.dooboo.product.dto.ProductimageDTO">
-		SELECT * FROM PRODUCT_IMAGE WHERE PRODUCT_NO = #{productNo}
-	</select>
 	
-	<select id="selectBoardList" parameterType="com.koreait.dooboo.communityboard.dto.CommunityBoardDTO" resultMap="boardResult">
 	
-			
+	<select id="selectBoardList" parameterType="com.koreait.dooboo.communityboard.dto.CommunityBoardDTO" resultType="com.koreait.dooboo.communityboard.dto.CommunityBoardDTO">
+	
   	 	<include refid="startPaging"/>
 	  		SELECT
 	  			ROWNUM,
@@ -171,8 +155,4 @@ PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
 				</if>
 			</if>
 	</select>
-	
-	<delete id="deleteFile" parameterType="com.koreait.dooboo.product.dto.ProductimageDTO">
-		DELETE FROM PRODUCT_IMAGE WHERE FILE_NO = #{fileNo} <!-- AND REG_NO = #{regNo} -->
-	</delete>	
 </mapper>
