+++ src/main/resources/mapper/product.xml
@@ -33,8 +33,25 @@ PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
 		, REG_DATE
 		, MOD_DATE
 	</sql>
+	
+	<resultMap id="productResult" type="com.koreait.dooboo.product.dto.ProductDTO">
+	    <result property="rnum" column="RNUM"/>
+	    <result property="productNo" column="PRODUCT_NO"/>
+	    <result property="title" column="TITLE"/>
+	    <result property="price" column="PRICE"/>
+	    <result property="hit" column="HIT"/>
+	    <result property="content" column="CONTENT"/>
+	    <result property="status" column="STATUS"/>
+	    <result property="address" column="ADDRESS"/>
+	    <result property="regNo" column="REG_NO"/>
+	    <result property="modNo" column="MOD_NO"/>
+	    <result property="regDate" column="REG_DATE"/>
+	    <result property="modDate" column="MOD_DATE"/>
+	    <result property="nickname" column="NICKNAME"/>
+	    <collection property="productImageList" column="PRODUCT_NO" javaType="java.util.ArrayList" ofType="com.koreait.dooboo.product.dto.ProductimageDTO" select="selectProductFileList"/>
+	</resultMap>
 
-<select id="selectProductList" parameterType="com.koreait.dooboo.product.dto.ProductDTO" resultType="com.koreait.dooboo.product.dto.ProductDTO">
+<select id="selectProductList" parameterType="com.koreait.dooboo.product.dto.ProductDTO" resultMap ="productResult">
 	
   	 	<include refid="startPaging"/>
 	  		SELECT
@@ -61,6 +78,7 @@ PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
 		<include refid="endPaging"/>				
 	</select>
 	
+	<!--  @@@@@@@@@@@@@@@@@@ 업로드한 이미지 파일을 관리한다     보람씨 작업한것 @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@-->
 	<update id="updateHit" parameterType="long">
 		UPDATE
 			PRODUCT 
@@ -79,7 +97,6 @@ PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
 			del_yn = 'N'
 	</select>
 
-	<!--  @@@@@@@@@@@@@@@@@@ 업로드한 이미지 파일을 관리한다     보람씨 작업한것 @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@-->
 	<insert id="insertProductImage" parameterType="com.koreait.dooboo.product.dto.ProductimageDTO">
 		INSERT INTO PRODUCT_IMAGE
 		(	
@@ -98,12 +115,18 @@ PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
 			, SYSDATE
 		)
 	</insert>
-<!-- @@@@@@@@@@@@@@@@@@@@@@@@@@@@@보람씨 작업한것@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@ -->
+	
+	
+	
+	<select id="selectProductFileList" resultType="com.koreait.dooboo.product.dto.ProductimageDTO">
+		SELECT * FROM PRODUCT_IMAGE WHERE PRODUCT_NO = #{productNo}
+	</select>
+	
+	<delete id="deleteFile" parameterType="com.koreait.dooboo.product.dto.ProductimageDTO">
+		DELETE FROM PRODUCT_IMAGE WHERE FILE_NO = #{fileNo} <!-- AND REG_NO = #{regNo} -->
+	</delete>	
+<!-- @@@@@@@@@@@@@@@@@@@@@@@@@@@@@보람씨 작업한것 END@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@ -->
 
-	
-	
-	
-	
 	<!-- 온석태 수정 -->
 		<!-- 판매상품을 등록한다 -->
 	<insert id="insertSellProduct">
