+++ src/main/webapp/WEB-INF/views/CommunityBoard/read.jsp
@@ -32,15 +32,6 @@
 		    	<label>내용</label>
 		    	<textarea name="content" id="inputGroup2_2" readonly>${board.getContent()}</textarea>
 		    </div>
-    		<div>
-		    	<label>첨부파일</label>
-		    	<c:forEach var="boardFile" items="${board.productImageList}" varStatus="status">
-		    		<div id="div_file_${boardFile.fileNo}">
-		    			<img src="${STATIC_IMAGE_ROOT }${boardFile.filePath }" width="100px">
-						<span>${boardFile.fileName}<button type="button" class="btn-close file-close" onclick="removeFile(this)" data-file-seq="${boardFile.fileNo}" aria-label="Close"></button></span>						
-		    		</div>
-		    	</c:forEach>
-		    </div>
 		  	<div>
 			    <label>등록일</label>
 			    <input name="regdate" value="${board.getRegDate()}" readonly >
