+++ src/main/webapp/WEB-INF/views/CommunityBoard/write.jsp
@@ -9,38 +9,25 @@
 <!-- End head -->
 <body>
 	<form id="frm" name="frm">
-       	<h1>커뮤니티 게시판</h1>
-	            <div>
-		         <div>
-                 <h3>게시글 작성하기</h3>
-	             <div>
-	                <label>Title</label>
-                    <input type="text"  value="${communityboard.getTitle()}" name="title" placeholder="제목을 입력해주세요">
-	             </div>
-                <div>
-                  	<label>Contents</label>
-                   	<textarea rows="3" name="content" placeholder="내용을 입력하세요"></textarea>
-                  
-                 </div>
-	    		<!-- <div>
-			    	<label>파일</label>
-			    	<input type="file" name="uploadFile" id="fileUpload" title="첨부파일 추가1">
-			    	<label>파일2</label>
-			    	<input type="file" name="uploadFile" id="fileUpload2" title="첨부파일 추가2">
-			    	<label>파일3</label>
-			    	<input type="file" name="uploadFile" id="fileUpload3" title="첨부파일 추가3">
-			    </div>         -->         
-                 <div style="text-align : center">
-	                <c:if test="${empty board}"> 
-	                		<a href="javascript:boardSave();">저장하기</a>
-					</c:if>
-					<c:if test="${not empty board}"> 
-	                		<a href="javascript:boardSave();">수정하기</a>
-					</c:if>
-	                	<button><a href="b.list">목록보기</a></button>
-                  <!-- End Textarea -->
-                </div>
-              </form>  
+      <h1>커뮤니티 게시판</h1>
+         <h3>게시글 작성하기</h3>
+           <div>
+              <label>Title</label>
+              <input type="text"  value="${communityboard.getTitle()}" name="title" placeholder="제목을 입력해주세요">
+           </div>
+           <div>
+           	 <label>Contents</label>
+           	 <textarea rows="3" name="content" placeholder="내용을 입력하세요"></textarea>
+           </div>     
+           <div style="text-align : center">
+             <c:if test="${empty board}"> 
+               <a href="javascript:boardSave();">저장하기</a>
+			 </c:if>
+			 <c:if test="${not empty board}"> 
+             	<a href="javascript:boardSave();">수정하기</a>
+			 </c:if>
+          <button><a href="b.list">목록보기</a></button>
+     </form>  
 	<script>
 		function boardSave(){
